Reorganize around LPC812 microcontroller
Previously we were just stubbing stuff out using an Adafruit Feather M0
board, but now we're getting ready to work with something resembling the
real target board. For the moment, we're using the mainboard from the
previous-generation miniclock as a victim development board, until the
new board is designed and brought up.

As of right now this is just an LPC812 blinking demo but the main focus
here is to get the build stuff set up correctly.