firmware: Push main crate down one level
When we're done we'll have a number of different crates here: one binary
crate for each supported target, and a set of shared library crates for
common functionality between the targets, such as the main application
logic.

This just gives us room to add new crates alongside; the initial firmware
crate is now named lpc812-mbed-xpresso reflecting the specific development
board it will initially target until we're ready to manufacture the "real"
PCB(s) for the project.