Rework the LPC812-MAX target's VSCode debug launch config
This makes it work at least part of the time, though it still doesn't seem
to work 100% reliably. It seems that when OpenOCD is first launched it
can fail to write to the code flash the first time it is asked to, but
then it tends to work on subsequent tries.