clockmain: Some more fonts
We're going to switch to using a tile-based display model, with 4x4 pixel
square tiles, in order to get the most out of the resolution of the OLED
display while still keeping the "chunky" aesthetic that the original big
4x4 super-pixel model was going for.

That requires some new fonts that are constructed from those tiles. These
fonts are given as normal bitmap fonts for ease of editing, but will be
processed into tile grids by a pre-processing program so that their
storage is more compact in the firmware.

For digitfont in particular, this one won't be stored as individual tiles
at all. Because the set of possibilities is limited by the 7-segment-like
presentation, we can pack all of the information needed to render a
character into a single 16-bit value that encodes which segments are
lit and how the segments are connected (curved or straight), which we
can then dynamically turn into an 8x16 tile grid at render time.