firmware/clockmain: The 16 big digit shapes
These are encoded in a compact way that exploits the 7-segment-style
presentation of these. For each digit we have a byte containing the
conventional 7-segment display segment encoding and then another byte
describing whether each of the segment intersections should have a curve
tile written into it. That gives us 16 bits per digit for a total of 32
bytes to encode all of the digit shapes.