firmware/simulator: Just inefficiently draw pixels, rather than textures
The libsdl2 texture management model is not really compatible with
creating trait-based abstractions on top of it because the lifetime
modeling of the textures requires every caller to track a special lifetime
for the object.

While I expect there is a clever way to get that working, the simulator
is just a tool to help with development of the main thing, so it's not
worth spending time fighting with Rust to optimize it when drawing pixels
individually already works well enough.